[systemd login] Handle ENODATA properly. Fixes JB#53297
Handle ENODATA similarly to ENOENT error. This is reported by systemd
when the data is not prepared yet. User change is monitored anyways, so
getting this error simply means that we do not quit the main loop just
yet and wait for the data to get ready to indicate what user is logged
in.